+++ test/unit/cross-instance.test.ts
@@ -8,7 +8,7 @@
 import assert from 'assert';
 import AsyncTestUtil, { wait, waitUntil } from 'async-test-util';
 
-import config from './config';
+import config, { getEncryptedStorage } from './config';
 import {
     isRxDatabase,
     createRxDatabase,
@@ -23,9 +23,6 @@ import * as schemas from './../helper/schemas';
 import * as schemaObjects from './../helper/schema-objects';
 import * as humansCollection from './../helper/humans-collection';
 import { HumanDocumentType } from './../helper/schemas';
-import {
-    wrappedKeyEncryptionStorage
-} from '../../plugins/encryption';
 
 config.parallel('cross-instance.test.js', () => {
     if (!config.storage.hasMultiInstance) {
@@ -173,18 +170,14 @@ config.parallel('cross-instance.test.js', () => {
             const password = randomCouchString(10);
             const db1 = await createRxDatabase({
                 name,
-                storage: wrappedKeyEncryptionStorage({
-                    storage: config.storage.getStorage(),
-                }),
+                storage: getEncryptedStorage(),
                 password,
                 multiInstance: true,
                 ignoreDuplicate: true
             });
             const db2 = await createRxDatabase({
                 name,
-                storage: wrappedKeyEncryptionStorage({
-                    storage: config.storage.getStorage(),
-                }),
+                storage: getEncryptedStorage(),
                 password,
                 multiInstance: true,
                 ignoreDuplicate: true
@@ -237,18 +230,14 @@ config.parallel('cross-instance.test.js', () => {
             const password = randomCouchString(10);
             const db1 = await createRxDatabase({
                 name,
-                storage: wrappedKeyEncryptionStorage({
-                    storage: config.storage.getStorage(),
-                }),
+                storage: getEncryptedStorage(),
                 password,
                 multiInstance: true,
                 ignoreDuplicate: true
             });
             const db2 = await createRxDatabase({
                 name,
-                storage: wrappedKeyEncryptionStorage({
-                    storage: config.storage.getStorage(),
-                }),
+                storage: getEncryptedStorage(),
                 password,
                 multiInstance: true,
                 ignoreDuplicate: true