+++ CHANGELOG.md
@@ -58,6 +58,7 @@
 - FIX `$regex` query not working on remote storage.
 - FIX SQLite must store attachments data as Blob instead of base64 string to reduce the database size.
 - FIX CouchDB replication conflict handling
+- CHANGE Encryption plugin was renamed to `encryption-crypto-js`
 
 #
 - REMOVE deprecated `babel-plugin-transform-async-to-promises` plugin.